feat(XSSF): updates shared strings handling for compatibility
